Use redirects and login instead of conditional split in the view layer
